+++ packages/components/radio/__tests__/radio.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import { Radio } from "../src";
+import { Radio } from "../src2";
 
 describe("Radio", () => {
   it("should render correctly", () => {