+++ packages/components/radio/src/index.ts
@@ -1,5 +1,2 @@
-"use client";
-
 export { Radio, type RadioProps } from "./radio";
-export { RadioGroup } from "./radio-group";
-export { RadioGroupItem, type RadioGroupItemProps } from "./radio-group-item";
+export { RadioGroup, type RadioGroupProps } from "./radio-group";