+++ packages/components/radio/src/radio-group-context.ts
@@ -1,0 +1,9 @@
+import { createContext } from "@jamsr-ui/utils";
+import type { ContextType } from "./use-radio-group";
+
+export const [RadioGroupProvider, useRadioGroupContext] = createContext<
+  ContextType | undefined
+>({
+  name: "RadioGroupContext",
+  strict: false,
+});